Create README.md
PROBLEM STATEMENT:
BSA cycles a famous bicycle company has collected the data of sharing of bicycle for two consecutive years. With this data they want to do an analysis on how the sharing of the bicycle happens in different environmental and seasonal changes. Thereby, they even want to predict the sharing of the data for the upcoming year. Hence, the goal of the study is to build a system and come up with the prediction for the following year.
OBJECTIVE:
To analyse different environmental conditions which are affecting the count of users for bicycle sharing.
To consider all the factors and come out with best features, which are related most with the users of bicycle sharing.
To give a clear view of the analysis to BSA Cycles, so that it can help the company to improve their bicycle sharing business.
To help BSA cycles with a prediction model based on the factors given for the upcoming year.
HYPOTHESIS:
Various hypothesis were taken into consideration while analysing each of the data features.